moved components inside of app
