+++ src/app/admin/QrCode/[projectName]/page.jsx
@@ -1,6 +1,6 @@
 'use server';
-import ClientComponent from "@/components/QrCode/clientComponent.jsx";
-import ServerComponent from "@/components/QrCode/serverComponent.jsx";
+import ClientComponent from "@/app/components/QrCode/clientComponent.jsx";
+import ServerComponent from "@/app/components/QrCode/serverComponent.jsx";
 export default async function QrCodePage({params}) {
   const siteName = params.projectName;
   return (
