+++ src/app/admin/decoy/page.jsx
@@ -1,5 +1,5 @@
 'use server';
-import ClientComponent from "@/components/adminDecoy/clientComponent";
+import ClientComponent from "@/app/components/adminDecoy/clientComponent";
 
 export default async function Home() {
