+++ src/app/admin/userData/page.jsx
@@ -1,7 +1,7 @@
 
 
-import UserDataDisplayServer from "@/components/userDataDisplay/serverComponent.jsx";
-import UserDataDisplayClient from "@/components/userDataDisplay/clientComponent.jsx";
+import UserDataDisplayServer from "@/app/components/userDataDisplay/serverComponent.jsx";
+import UserDataDisplayClient from "@/app/components/userDataDisplay/clientComponent.jsx";
 
 export default async function userData () {
   return (
