+++ src/app/components/QrCode/QrCode.module.css
