+++ src/app/components/adminDashboard/clientComponent.jsx
@@ -1,7 +1,7 @@
 //@/components/adminDashboard/clientComponent.jsx
 'use client';
 import { useRouter } from "next/navigation";
-import adminStyles from "@/components/adminDashboard/adminDashboard.module.css";
+import adminStyles from "@/app/components/adminDashboard/adminDashboard.module.css";
 import Button from "@mui/material/Button";
 import { logout } from "@/actions/actions.jsx";
 // import ClientButtonToQrCode from '@/components/adminDashboard/clientButtonToQrCode.jsx';
