+++ src/app/components/adminDecoy/clientComponent.jsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/navigation';
 
 import Button from "@mui/material/Button";
 import { TextField } from "@mui/material";
-import adminStyles from "@/components/adminDecoy/adminDecoy.module.css";
+import adminStyles from "@/app/components/adminDecoy/adminDecoy.module.css";
 import { socket } from "@/app/socket.js";
 import { getWebSites } from "@/actions/actions.jsx";
 export default function Home() {
