+++ src/app/components/login/password.jsx
@@ -4,8 +4,8 @@ import { TextField } from "@mui/material";
 import Button from "@mui/material/Button";
 import Link from "next/link";
 import Image from "next/image";
-import notGoogleImage from "../../../public/notGoogle.png";
-import adminImage from "../../../public/admin.png";
+import notGoogleImage from '/public/notGoogle.png';
+import adminImage from "/public/admin.png";
 
 import notGoogleStyles from "./notGoogleStyles.module.css";
 import testcss from "./testcss.module.css";
