+++ src/app/page.jsx
@@ -1,5 +1,5 @@
 "use client";
-import Login from "@/components/login/login";
+import Login from "@/app/components/login/login";
 
 export default function Home() {  
   return (
